import in plot and test corrected. test moved to test_da folder
